refactor: 로또 게임 재시작 상수 view 로 이동
