+++ __tests__/LottoGame.test.js
@@ -40,21 +40,4 @@ describe("로도 판매 기계 기능 테스트", () => {
       expect(validateAvailableLottoCount).toThrow(errorMessage);
     }
   );
-
-  test.each(["123", "N", -1])(
-    "로또 게임 재시작 여부가 y 또는 n이 아닌 경우 에러가 발생한다.",
-    (input) => {
-      // given
-
-      // when
-      const validate = () => {
-        LottoGame.validateIsRestartLottoGame(input);
-      };
-
-      // then
-      expect(validate).toThrow(
-        ErrorLottoGame.ERROR_LOTTO_GAME_RESTART_NOT_VALID
-      );
-    }
-  );
 });
