+++ src/js/view/console/Input.js
@@ -1,6 +1,10 @@
 import { readLineAsync } from "../../../utils/readlineAsync";
+import { ErrorLottoGameRestart } from "../../constants/error";
 
 const Input = {
+  RESTART_GAME_TRUE: "y",
+  RESTART_GAME_FALSE: "n",
+
   async getLottoPurchasedAmount() {
     return await readLineAsync("구입금액을 입력해 주세요.");
   },
@@ -14,7 +18,17 @@ const Input = {
   },
 
   async getIsRestartLottoGame() {
-    return await readLineAsync("\n다시 시작하시겠습니까? (y/n)");
+    const input = await readLineAsync("\n다시 시작하시겠습니까? (y/n)");
+
+    this.validateIsRestartLottoGame(input);
+
+    return input === Input.RESTART_GAME_TRUE;
+  },
+
+  validateIsRestartLottoGame(input) {
+    if (![this.RESTART_GAME_FALSE, this.RESTART_GAME_TRUE].includes(input)) {
+      throw new Error(ErrorLottoGameRestart.ERROR_LOTTO_GAME_RESTART_NOT_VALID);
+    }
   },
 };
